RESTful API/requirements.txt
Flask==2.0.1
google-cloud-datastore==2.4
requests~=2.27.1